requirements.txt
fastapi==0.100.0
pydantic==2.0.2
uvicorn==0.22.0
torch>=2.0.0
transformers>=4.34.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-multipart==0.0.6
typing-extensions==4.5.0
numpy==1.24.3
python-dotenv==1.0.0
aiohttp==3.9.5
openai==0.27.8
langchain==0.0.228
accelerate==0.26.0
bitsandbytes==0.41.1
sentencepiece
protobuf
safetensors
